reachable-functions now considers makeInterface
The reachable-functions analysis now considers methods that are invoked via
an interface that is created via a makeInterface instruction.  The analysis
computes the set of methods required for the interface, and then filters the
set of methods given as argument to makeInterface using this set.  All
resulting methods are considered reachable.

cr: https://code.amazon.com/reviews/CR-54478980