Fix function thunk summaries * Fix function thunk summaries
The taint analysis was not reporting function "thunk" callee summaries
because they are represented differently from "regular"
functions (similar *ssa.Function data but different pointers and names).

The fix was to compare the callee and summarized function name prefixes
and positions to see if they are in fact the same.